fix(#1): add inheritance of stdin, add issue tests
The library uses `std::process::Command::output()` function for creating
a new instance of `tmux` and invoking it's subcommands (`new-session`,
`attach-session`, etc). `Stderr` and `Stdout` will be captured and
`Stdin` is not inherited from the parent process, and will be closed
after any attemp of child process to read from it, accordingly to the
documentation
[std/process/struct.Command](https://doc.rust-lang.org/std/process/struct.Command.html#method.output)
It causes `tmux` exit with error reporting `open terminal failed: not a
terminal`. To avoid this `Stdin` is set to `Stdio::inherit()` too.

Fixes: #1